Bugfix nullpointer ShowMapFragment.setupNmeaClientService / Cache offline received NMEA messages
Bugfix ShowMapFragment.setupNmeaClientService:
  Only get/use Activity when the Fragment isAdded.

When receiving NMEA messages while not connected to network,
cache these messages in a file at the client.
Connection to network is checked only once when starting application.
When restarting application while connected to Network,
these messages will be send to the backend (SocketIO).
Sending is only done when the Use setting allows this.