Bugfix: CalibrateTask throwing IllegalArgumentException when unbinding
Solving exception:
IllegalArgumentException (@CalibrateTask:destroyNmeaUdpClientService:122) {CalibrateTask}

The context.unbindService could potentially be called before the service has actually been bound.
The 'isBound'  boolean indicates that the context.bindService() was called and queued up for main thread execution.
When this is the case (isBound is true), it is allowed to call context.unbindService (which is queued to be executed).

More info:
https://stackoverflow.com/questions/22079909/android-java-lang-illegalargumentexception-service-not-registered